+++ OrangeDesignSystem/Sources/OrangeDesignSystem/Components/TextField/ODSTextFieldStyle.swift
@@ -24,31 +24,41 @@
 import SwiftUI
 
 ///
-/// Simple button with icon.
+/// <a href="https://system.design.orange.com/0c1af118d/p/47d389-text-fields/b/461794" target="_blank">ODS Text Field</a>.
 ///
-public struct ODSIconButton: View {
+/// The text field component comprises the text field itself, text selection and the edit menu. Some elements are styled and some are native.
+///
+
+extension View {
+    /// Sets the ods style on __TextField__ and __TextEditor__
+    public func odsTextFieldStyle() -> some View {
+        modifier(ODSTextFieldStyle())
+    }
+}
+
+// MARK: - Internal font modifier
+
+///
+/// Private modifier to get the theme in environment.
+///
+
+private struct ODSTextFieldStyle: ViewModifier {
+
+    // =======================
+    // MARK: Stored Properties
+    // =======================
+
     @Environment(\.theme) private var theme
 
-    let image: Image
-    let action: () -> Void
+    // ==========
+    // MARK: Body
+    // ==========
 
-    /// Initialize the button.
-    ///
-    /// - Parameters:
-    ///   - image: The icon to be displayed.
-    ///   - action: Will be called when the user clicks the button.
-    ///
-    public init(image: Image, action: @escaping () -> Void) {
-        self.image = image
-        self.action = action
-    }
-
-    public var body: some View {
-        Button {
-            action()
-        } label: {
-            ODSIcon(image)
-                .foregroundColor(theme.componentColors.accent)
-        }
+    func body(content: Content) -> some View {
+        content
+            .accentColor(theme.componentColors.accent)
+            .padding(.all, ODSSpacing.s)
+            .odsFont(.bodyRegular)
+            .background(Color(.tertiarySystemFill), in: RoundedRectangle(cornerRadius: 8.0))
     }
 }