+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.8.0] - 2022-11-30
+
+- [SDK] #295 Add NOTICE.txt file for Orange Images usage 
+- [DemoApp] #260 Update the Bar Tab variant list selection 
+- [SDK] Bug #276,#275,#279 [Accessibility] Do not vocalize images in ODSSmallCard, ODSCardTitleFirst and ODSCardImageFirst
+- [DemoApp] #259 Update the Bar Navigation variant list selection 
+- [DemoApp/SDK] #51 Add banners component
+- [SDK] #289 Add documentation for bars-tab component
+- [DemoApp] #257 Guidelines - Typography text description updates 
+- [DemoApp/SDK] #267 Update the list of Text fields variants
+- [DemoApp] #258 Add the play icon to the component's variant lists
+
 ## [0.7.0] - 2022-11-07
 
 - [DemoApp/SDK] #113 Add theme management and create Orange and InnovationCup themes
@@ -14,7 +26,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - [DemoApp] #248 Technical code refactoring for component page presentation 
 - [SDK/DemoApp] #229 Technical update to use Small and ImageFirst cards in demo app and modules 
 - [DemoApp] #228 Change Chip to switch on List screen for "SecondLineOfText" and "LeadingImage" options 
-- [SDK/DemoApp] #237 Define a Grid of small cards module and use it in application for components
 - [SDK/DemoApp] #182 Add card title first
 - [SDK/DemoApp] #237 Define a Grid of small cards module and use it in application for components 
 - [DemoApp] #235 Rename Module called Card View to Card Collections (2)